# api/openapi.yaml
openapi: "3.0.3"
info:
  title: CMDB Platform API
  version: "1.0.0"
  description: Configuration Management Database Platform API
servers:
  - url: /api/v1
security:
  - BearerAuth: []
...
paths:
...
  /auth/change-password:
    post:
      operationId: changePassword
      tags: [auth]
      summary: Change the authenticated user's password
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [current_password, new_password]
              properties:
                current_password:
                  type: string
                new_password:
                  type: string
      responses:
        '200':
          description: Password changed
        '400':
          $ref: '#/components/responses/BadRequest'

  # ── Topology ────────────────────────────────────────────────────────────
...
  /topology/dependencies:
    get:
      operationId: listAssetDependencies
      tags: [topology]
      summary: List asset dependencies
      parameters:
        - name: asset_id
          in: query
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Dependency list
    post:
      operationId: createAssetDependency
      tags: [topology]
      summary: Create asset dependency
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
      responses:
        '201':
          description: Dependency created
  /topology/dependencies/{id}:
    delete:
      operationId: deleteAssetDependency
      tags: [topology]
      summary: Delete asset dependency
      parameters:
        - $ref: '#/components/parameters/IdPath'
      responses:
        '204':
          description: Dependency deleted
  /topology/graph:
    get:
      operationId: getTopologyGraph
      tags: [topology]
      summary: Get topology graph for a location
      parameters:
        - name: location_id
          in: query
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Topology graph data

  # ── Assets ──────────────────────────────────────────────────────────────
...
  /users/{id}/sessions:
    get:
      operationId: listUserSessions
      tags: [identity]
      summary: List active sessions for a user
      parameters:
        - $ref: '#/components/parameters/IdPath'
      responses:
        '200':
          description: Session list

  # ── Prediction ──────────────────────────────────────────────────────────
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  parameters:
    IdPath:
      name: id
      in: path
      required: true
      schema:
        type: string
        format: uuid
...
  responses:
    BadRequest:
      description: Bad request
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
...
  schemas:
...
    Meta:
      type: object
      properties:
        request_id:
          type: string
      required: [request_id]
...
    ErrorBody:
      type: object
      properties:
        code:
          type: string
        message:
          type: string
      required: [code, message]
    ErrorResponse:
      type: object
      properties:
        error:
          $ref: '#/components/schemas/ErrorBody'
        meta:
          $ref: '#/components/schemas/Meta'
      required: [error, meta]

    # ── Auth ──────────────────────────────────────────────────────────────
...